修改试卷分类，试题分类卡控，试卷递归呢查找

+++ ruoyi-exam/src/main/java/com/ruoyi/exam/controller/ExamPaperCategoryController.java
@@ -5,6 +5,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
+import com.ruoyi.exam.domain.ExamPaper;
+import com.ruoyi.exam.service.IExamPaperService;
 import com.ruoyi.framework.web.util.ShiroUtils;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,6 +40,9 @@ public class ExamPaperCategoryController extends BaseController
 	
 	@Autowired
 	private IExamPaperCategoryService examPaperCategoryService;
+
+	@Autowired
+	private IExamPaperService examPaperService;
 	
 	@RequiresPermissions("exam:examPaperCategory:view")
 	@GetMapping()
@@ -57,8 +62,8 @@ public class ExamPaperCategoryController extends BaseController
         List<ExamPaperCategory> list = examPaperCategoryService.selectExamPaperCategoryPage(examPaperCategory);
 		return list;
 	}
-	
-	
+
+
 	/**
 	 * 导出试卷分类列表
 	 */
@@ -91,6 +96,11 @@ public class ExamPaperCategoryController extends BaseController
 	@ResponseBody
 	public AjaxResult addSave(ExamPaperCategory examPaperCategory)
 	{
+		ExamPaper examPaper = new ExamPaper();
+		examPaper.setExamPaperCategoryId(examPaperCategory.getParentId());
+		if(examPaperService.selectList(examPaper).size()>0){
+			return error(1, "分类包含试卷，不允许扩展分类");
+		}
 		examPaperCategory.setCreateBy(ShiroUtils.getLoginName());
 		examPaperCategory.setCreateDate(new Date());
 		examPaperCategory.setDelFlag("0");
@@ -104,7 +114,9 @@ public class ExamPaperCategoryController extends BaseController
 	public String edit(@PathVariable("id") Integer id, ModelMap mmap)
 	{
 		ExamPaperCategory examPaperCategory = examPaperCategoryService.selectById(id);
+		ExamPaperCategory parent = examPaperCategoryService.selectById(examPaperCategory.getParentId());
 		mmap.put("examPaperCategory", examPaperCategory);
+		mmap.put("parentName", parent.getName());
 	    return prefix + "/edit";
 	}
 	
@@ -119,7 +131,21 @@ public class ExamPaperCategoryController extends BaseController
 	{
 		examPaperCategory.setUpdateBy(ShiroUtils.getLoginName());
 		examPaperCategory.setUpdateDate(new Date());
-		return toAjax(examPaperCategoryService.updateSelectiveById(examPaperCategory));
+
+		ExamPaper examPaper = new ExamPaper();
+		examPaper.setExamPaperCategoryId(examPaperCategory.getParentId());
+		if(examPaperService.selectList(examPaper).size()>0){
+			return error(1,"父级分类下面包含试卷，无法移动");
+		}
+
+		ExamPaperCategory exam = new ExamPaperCategory();
+		exam.setParentId(examPaperCategory.getId());
+		if (examPaperCategoryService.selectList(exam).size() > 0)
+		{
+			return error(1, "此分类存在下级分类,无法移动");
+		}
+
+			return toAjax(examPaperCategoryService.updateSelectiveById(examPaperCategory));
 	}
 	
 	/**
@@ -151,16 +177,8 @@ public class ExamPaperCategoryController extends BaseController
 	public List<Map<String, Object>> treeDataForAdd()
 	{
 		List<Map<String, Object>> tree = examPaperCategoryService.selectDeptTree();
-		List<Map<String, Object>> res = new ArrayList<>();
-		for (Map<String, Object> stringObjectMap : tree) {
-			String pId = stringObjectMap.get("pId").toString();
-			if(pId.equals("0")||pId.equals("1")){
-				res.add(stringObjectMap);
-			}
 
-		}
-
-		return res;
+		return tree;
 	}
 
 	/**
@@ -178,6 +196,12 @@ public class ExamPaperCategoryController extends BaseController
 		{
 			return error(1, "存在下级分类,不允许删除");
 		}
+		ExamPaper examPaper = new ExamPaper();
+		examPaper.setExamPaperCategoryId(id);
+		if(examPaperService.selectList(examPaper).size()>0){
+			return error(1, "分类包含试卷，不允许删除分类");
+		}
+
 
 		return toAjax(examPaperCategoryService.deleteById(id));
 	}
